Add unprefixed transform & backface-visibility CSS
The slide hiding/showing in this presentation depends on these CSS classes, which currently only use webkit-prefixed CSS and are missing the standard unprefixed versions. So, the presentation doesn't work at all in Firefox 46 (or any other non-webkit-derived browser that doesn't include webkit prefixes for compatibility). All the slides are shown stacked on top of each other, in a jumble of letters.

This pull request adds the unprefixed syntax, which (based on local testing) fixes the issue in Firefox at least.